Added src and build directories
Currently, because the program uses the absolute file path of the
calling terminal, the program should now be run through the run.sh file
for convinences sake.

This is a temporary solution, in a later commit the source will be
altered to work out the absolute path of the executable on its own.